Rearrange mapper test directory
