+++ test/integration/short/mapper/mapper-batch-tests.js
@@ -1,0 +1,16 @@
+'use strict';
+
+const mapperTestHelper = require('./mapper-test-helper');
+
+
+describe('Mapper', function () {
+
+  mapperTestHelper.setupOnce(this);
+
+  describe('#batch()', () => {
+
+    it('should execute a batch containing multiple tables from the same doc', () => {
+
+    });
+  });
+});
